requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
django==1.8.9
six==1.10.0
djangorestframework==3.3.2
django-autofixture==0.12.1



